requirements/common.txt
alembic==0.8.8
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.0
APScheduler==3.3.0
billiard==3.3.0.23
blinker==1.4
business-calendar==0.2.1
cached-property==1.3.0
celery==3.1.23
cffi==1.9.1
click==6.6
coverage==4.2
cryptography==1.7.2
dominate==2.2.1
elasticsearch==5.0.1
Flask==0.11.1
Flask-APScheduler==1.5.0
Flask-Bootstrap==3.3.7.0
Flask-Elasticsearch==0.2.5
Flask-KVSession==0.6.2
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.0
Flask-Moment==0.5.1
Flask-ReCaptcha==0.4.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Tracy==0.1.3
Flask-WTF==0.12
holidays==0.5
idna==2.2
isodate==0.5.4
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.6.0
kombu==3.0.35
ldap3==2.1.1
Mako==1.0.4
MarkupSafe==0.23
oauthlib==2.0.1
packaging==16.8
paramiko==2.0.2
pyasn1==0.2.2
pycparser==2.17
pycrypto==2.6.1
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.5.3
python-dotenv==0.6.0
python-editor==1.0.1
python-magic==0.4.12
pytz==2016.6.1
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.7.0
simplekv==0.10.0
six==1.10.0
SQLAlchemy==1.1.4
tzlocal==1.3
urllib3==1.20
virtualenv==13.1.2
visitor==0.1.3
Werkzeug==0.11.11
WTForms==2.1
